2001-11-19 Dan Mills <dmills@spamblock.demon.co.uk>
	* fit.c (split_at_corners):
	 Here is a patch for another autotrace bug which caused some images
	 to seemingly hang when processed with the thin option.

	 The problem was caused by the fact that the LHS of the if statement
	 was evaluated as an unsigned value, thus when pixel_o == 1, the
	 statement evaluates as if (((unsigned int)(-1)) >= 4) which comes
	 out to somewhere around (2^32 >= 4), this has the predictable result.
	 Using signed math overcomes this problem.

2001-11-19  Martin Weber  <martweb@gmx.net>

	* pxl-outline.c (find_outline_pixels): improved
	background handling.

+++ fit.c
@@ -394,8 +394,8 @@ split_at_corners (pixel_outline_list_type pixel_list, fitting_opts_type *fitting
       if (O_LENGTH (pixel_o) > fitting_opts->corner_surround * 2 + 2)
         corner_list = find_corners (pixel_o, fitting_opts);
       else {
-        unsigned int surround;
-        if ((surround = (O_LENGTH (pixel_o) - 3) / 2) >= 4)
+        int surround;
+        if ((surround = (int)(O_LENGTH (pixel_o) - 3) / 2) >= 4)
           {
             unsigned save_corner_surround = fitting_opts->corner_surround;
             fitting_opts->corner_surround = surround;
